Made a separate service for db logic
